Auth done

+++ server/controller/posts.js
@@ -5,22 +5,24 @@ import PostMessage from "../models/postMessage.js";
 export const getPosts = async (req, res) => {
   try {
     const postMessages = await PostMessage.find();
-    res.status(200).json(postMessages);
+    return res.status(200).json(postMessages);
   } catch (error) {
-    res.status(404).json({
+    return res.status(404).json({
       message: error.message,
     });
   }
 };
 
 export const createPosts = async (req, res) => {
-  const post = req.body;
+  let post = req.body;
+  const userId = req.userId;
+  post = { ...post, creatorId: userId, createdAt: new Date().toISOString() };
   const newPost = new PostMessage(post);
   try {
     await newPost.save();
-    res.status(201).json(newPost);
+    return res.status(201).json(newPost);
   } catch (error) {
-    res.status(409).json({
+    return res.status(409).json({
       message: error.message,
     });
   }
@@ -29,8 +31,14 @@ export const createPosts = async (req, res) => {
 export const updatePost = async (req, res) => {
   const { id: _id } = req.params;
   const post = req.body;
+
   if (!mongoose.Types.ObjectId.isValid(_id))
     return res.status(404).send("No Post with that Id");
+
+  const foundPost = await PostMessage.findById(_id);
+
+  if (req.userId !== foundPost.creatorId)
+    return res.status(404).json({ message: "User did not create this post" });
 
   const updatedPost = await PostMessage.findByIdAndUpdate(
     _id,
@@ -40,15 +48,19 @@ export const updatePost = async (req, res) => {
     }
   );
 
-  res.json(updatedPost);
+  return res.json(updatedPost);
 };
 
 export const deletePost = async (req, res) => {
   const { id: _id } = req.params;
   if (!mongoose.Types.ObjectId.isValid(_id))
     return res.status(404).send("No Post with that Id");
+  const post = await PostMessage.findById(_id);
+  console.log(post);
+  if (req.userId !== post.creatorId)
+    return res.status(400).json({ message: "User did not create this post" });
   await PostMessage.findByIdAndRemove(_id);
-  res.json({ message: "Post deleted Successfully" });
+  return res.json({ message: "Post deleted Successfully" });
 };
 
 export const likePost = async (req, res) => {
@@ -68,5 +80,5 @@ export const likePost = async (req, res) => {
   const updatedPost = await PostMessage.findByIdAndUpdate(_id, post, {
     new: true,
   });
-  res.json(updatedPost);
+  return res.json(updatedPost);
 };
